guest video fixed
- better video
- added texts for response

+++ code/app/src/main/java/de/dhkarlsruhe/it/sheeshapp/sheeshapp/guest/LoginActivityGuest.java
@@ -5,7 +5,9 @@ import android.support.design.widget.Snackbar;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.View;
+import android.widget.Button;
 import android.widget.EditText;
+import android.widget.TextView;
 
 import com.google.android.gms.ads.AdRequest;
 import com.google.android.gms.ads.MobileAds;
@@ -25,6 +27,9 @@ public class LoginActivityGuest extends AppCompatActivity implements RewardedVid
     private final static String AD_REWARD_ID_TEST = "ca-app-pub-3940256099942544/5224354917";
     private RewardedVideoAd mRewardedVideoAd;
     private String username;
+    private boolean rewared = false;
+    private Button btnLogin;
+    private TextView tvInfo;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -32,6 +37,8 @@ public class LoginActivityGuest extends AppCompatActivity implements RewardedVid
         setContentView(R.layout.guest_activity_login);
         et = findViewById(R.id.etGuestLogin);
         guest = new Guest(this);
+        btnLogin = findViewById(R.id.btnGuestLogin);
+        tvInfo = findViewById(R.id.tvGuestLoginInfo);
         MobileAds.initialize(this,AD_APP_ID);
         mRewardedVideoAd = MobileAds.getRewardedVideoAdInstance(this);
         mRewardedVideoAd.setRewardedVideoAdListener(this);
@@ -39,8 +46,8 @@ public class LoginActivityGuest extends AppCompatActivity implements RewardedVid
     }
 
     private void loadRewardedVideoAd() {
-
-        mRewardedVideoAd.loadAd(AD_REWARD_ID, new AdRequest.Builder().build());
+        mRewardedVideoAd.loadAd(AD_REWARD_ID_TEST, new AdRequest.Builder().build());
+        tvInfo.setText(getString(R.string.loading_ad_text));
     }
 
 
@@ -54,27 +61,31 @@ public class LoginActivityGuest extends AppCompatActivity implements RewardedVid
             et.setText("");
             Snackbar.make(view,getString(R.string.check_your_inout),Snackbar.LENGTH_LONG).show();
         } else {
+            btnLogin.setEnabled(false);
             username = et.getText().toString().trim();
-            startAd();
+            startAd(view);
         }
     }
 
-    private void startAd() {
+    private void startAd(View view) {
         if (mRewardedVideoAd.isLoaded()) {
             mRewardedVideoAd.show();
         } else {
+            /*
             guest.deleteAll();
             guest.setName(username);
             guest.setEmail(username);
             Intent intent = new Intent(this, MainActivityGuest.class);
             startActivity(intent);
-            this.finish();
+            this.finish();*/
+            Snackbar.make(view,getString(R.string.error_text),Snackbar.LENGTH_LONG).show();
         }
     }
 
     @Override
     public void onRewardedVideoAdLoaded() {
-
+        btnLogin.setEnabled(true);
+        tvInfo.setText(getString(R.string.guest_login_info_text));
     }
 
     @Override
@@ -90,16 +101,21 @@ public class LoginActivityGuest extends AppCompatActivity implements RewardedVid
     @Override
     public void onRewardedVideoAdClosed() {
         loadRewardedVideoAd();
+        if (rewared) {
+            guest.deleteAll();
+            guest.setName(username);
+            guest.setEmail(username);
+            Intent intent = new Intent(this, MainActivityGuest.class);
+            startActivity(intent);
+            this.finish();
+        } else {
+            Snackbar.make(btnLogin, R.string.clsoed_ad_text, Snackbar.LENGTH_LONG).show();
+        }
     }
 
     @Override
     public void onRewarded(RewardItem rewardItem) {
-        guest.deleteAll();
-        guest.setName(username);
-        guest.setEmail(username);
-        Intent intent = new Intent(this, MainActivityGuest.class);
-        startActivity(intent);
-        this.finish();
+        rewared = true;
     }
 
     @Override
@@ -109,6 +125,8 @@ public class LoginActivityGuest extends AppCompatActivity implements RewardedVid
 
     @Override
     public void onRewardedVideoAdFailedToLoad(int i) {
+        Snackbar.make(btnLogin, getString(R.string.failed_to_load_ad), Snackbar.LENGTH_LONG).show();
+        tvInfo.setText(getString(R.string.failed_to_load_ad));
 
     }
 
@@ -116,6 +134,7 @@ public class LoginActivityGuest extends AppCompatActivity implements RewardedVid
     public void onRewardedVideoCompleted() {
 
     }
+
     @Override
     public void onResume() {
         mRewardedVideoAd.resume(this);
